+++ src/gantry_control/gui_server/action/__init__.py
@@ -6,57 +6,15 @@ need to be called after the GUISession object is initialized
 
 """
 
-import gmqclient
-
-from ..cli.format import _timestamp_
-from ..cli.progress_monitor import session_iterate
-from .session import (ActionCode, ActionEntry, ActionStatus,  # For typing
-                      GUISession)
-from .sync_socket import (sync_action_append, sync_action_status_update,
-                          sync_board_status, sync_full_session,
-                          sync_hardware_status)
-
-
-# Actual methods to processing
-def test_single_shot(session: GUISession, line):
-    """This is a simple test"""
-    session.logger.info("Running single shot testing!!")
-    for char in session_iterate(session, line):
-        session.logger.warn(f"Got characters {char}")
-        session.sleep(1)
-
-
-def gmq_disconnect(session: GUISession):
-    """"""
-    session.logger.info("Disconnected from GMQ server")
-    if session.hw is not None:
-        session.hw.close()
-    session.hw = None
-    sync_hardware_status(session)
-
-
-def gmq_connect(session: GUISession, host: str, port: int):
-    """Connection to Gantry MQ system"""
-    session.logger.info(f"Attempting to connect to GMQ server {host}:{port}")
-    if session.hw is not None:
-        session.hw.close()
-    session.hw = gmqclient.create_default_client(host, port)
-    session.hw.claim_operator()
-    sync_hardware_status(session)
-
-
-def gantry_move_to(session: GUISession, x: float, y: float, z: float):
-    session.hw.move_to(x=x, y=y, z=z)
-
-
-def start_new_session(session: GUISession, board_type: str, board_id: str):
-    session._init_board(f"++{board_type}@{board_id}")
-    sync_board_status(session)
+from ...cli.format import _timestamp_
+from ..session import ActionCode, ActionEntry, ActionStatus, GUISession
+from ..sync import action, sync_full_session
+from . import board, hardware, test_action
 
 
 # Main methods to keep track of the client-side requested action.
-def start_action(session: GUISession, name, **kwargs):
-    sync_action_append(
+def _start_action(session: GUISession, name, **kwargs):
+    action.sync_action_append(
         session,
         ActionEntry(
             name=name,
@@ -70,8 +28,8 @@ def start_action(session: GUISession, name, **kwargs):
     )
 
 
-def complete_action(session: GUISession, status=ActionCode.COMPLETE, **kwargs):
-    sync_action_status_update(
+def _complete_action(session: GUISession, status=ActionCode.COMPLETE, **kwargs):
+    action.sync_action_status_update(
         session, ActionStatus(timestamp=_timestamp_(), status=status, **kwargs)
     )
 
@@ -85,14 +43,14 @@ def halt_from_gui_user(session: GUISession):
 # Main methods to be exposed via the socket interface
 __run_action_method_map__ = {
     # Testing actions (should probably be removed for production)
-    "single-shot-test": test_single_shot,
+    "single-shot-test": test_action.test_single_shot,
     # Connecting to the various hardware controller clients
-    "gmq_disconnect": gmq_disconnect,
-    "gmq_connect": gmq_connect,
+    "gmq_disconnect": hardware.gmq_disconnect,
+    "gmq_connect": hardware.gmq_connect,
     # Simple control instructions
-    "gantry_move_to": gantry_move_to,
-    # Starting a new session 
-    "start-new-session": start_new_session,
+    "gantry_move_to": hardware.gantry_move_to,
+    # Starting a new session
+    "start-new-session": board.start_new_session,
 }
 
 
@@ -120,7 +78,7 @@ def register_action_sockets(session: GUISession):
 
         action_name = msg["name"]
         action_args = msg["args"]
-        start_action(session, action_name, args=action_args)
+        _start_action(session, action_name, args=action_args)
         return_status = ActionCode.COMPLETE
         # Unlocking the session again when starting a new command
         session._user_interupt = False
@@ -142,7 +100,7 @@ def register_action_sockets(session: GUISession):
         finally:
             session.logger.info(f"Completing action [{action_name}]")
             session._user_interupt = False  # Always release!!
-            complete_action(session, status=return_status)
+            _complete_action(session, status=return_status)
 
     @session.socket.on("user-interupt")
     def user_interupt():
