--- src/gui_client/src/Components/Tile.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-type Props = {};
-
-const Tile = (props: Props) => {
-  return <div>Tile</div>;
-};
-
-export default Tile;
